+++ src/scripts/village/reducers/fakeServer/firstMorning.ts
